% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_plot.R
\name{create_hex_plot}
\alias{create_hex_plot}
\title{Create a hexbin plot}
\usage{
create_hex_plot(
  data,
  x,
  y,
  bins = 20,
  title = NULL,
  x_label = NULL,
  y_label = NULL,
  plot_theme = theme_bw()
)
}
\arguments{
\item{data}{A data frame containing at least two numeric variables.}

\item{x}{The numeric variable on x-axis}

\item{y}{The numeric variable on y-axis}

\item{bins}{The number of hexagonal bins}

\item{title}{The plot title.}

\item{x_label}{The x-axis label}

\item{y_label}{The y-axis label}

\item{plot_theme}{The ggplot theme to use. The default is theme_bw()}
}
\value{
A ggplot object showing a hexbin plot.
}
\description{
This function creates a hexbin plot for any two numeric variables in a data
frame.
}
\examples{
cases <- load_data()

create_hex_plot(
  data = cases,
  x = year,
  y = measles_total)
}
