% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_country.R
\name{filter_by_country}
\alias{filter_by_country}
\title{Filter measles dataset by country}
\usage{
filter_by_country(country = "Madagascar", by = "cases_year")
}
\arguments{
\item{string}{dataset to be filtered (year or month)}
}
\value{
tibble
}
\description{
Filter measles dataset by country
}
\examples{
filter_by_country(country = "United States of America", by = "cases_month")

}
