% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_measles_gls.R
\name{fit_measles_gls}
\alias{fit_measles_gls}
\title{Fit the measles GLS model}
\usage{
fit_measles_gls(data)
}
\arguments{
\item{data}{The final measles dataset}
}
\value{
A GLS model object
}
\description{
This function fits the generalized least squares model used in the measles
analysis. It creates all the variables used in the model and accounts for
correlation.
}
