% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_plot_label.R
\name{hex_plot_label}
\alias{hex_plot_label}
\title{Make a nice label for a hex plot}
\usage{
hex_plot_label(var)
}
\arguments{
\item{var}{A variable name or expression.}
}
\value{
A character string formatted as a plot label.
}
\description{
This is a helper function that goes along with the hex_plot function.
This converts the variable names into nice labels for the plot.
}
