% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab-coverage.R
\name{lab_coverage}
\alias{lab_coverage}
\title{Get top or bottom 10 countries by proportion of lab confirmed to total measles cases}
\usage{
lab_coverage(year, top = TRUE)
}
\arguments{
\item{year}{A single numeric year between 2012 and 2025.}

\item{top}{Logical. If \code{TRUE} (default), returns the top 10 countries with
the highest lab confirmed:total cases ratio and the top 10 countries by total cases.
If \code{FALSE}, returns the bottom 10 countries.}
}
\value{
Prints a summary message and a tibble of 10 countries with their
lab confirmed:total proportion.
}
\description{
Filters the monthly cases dataset to a given year, aggregates lab confirmed, total measles
cases by country, and returns the top or bottom 10 countries proportions.
}
\examples{
lab_coverage(2019)
lab_coverage(2019, top = FALSE)
}
