% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_analysis.R
\name{linear_analysis}
\alias{linear_analysis}
\title{Fit a linear regression of log-average measles cases on Air Quality Index and runs a statistical analysis on the model}
\usage{
linear_analysis()
}
\value{
A \code{gt} table object with linear model analysis
}
\description{
Fit a linear regression of log-average measles cases on Air Quality Index and runs a statistical analysis on the model
}
\examples{
aqi_regression_table()
}
