% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load data set of user's choice (helper function)}
\usage{
load_data(file = "cases_month")
}
\arguments{
\item{string}{\code{cases_month}, \code{cases_year}, or \code{airquality} file to be loaded}
}
\value{
A tibble containing measles data
}
\description{
Load data set of user's choice (helper function)
}
