% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_detection_methods.R
\name{plot_detection_methods}
\alias{plot_detection_methods}
\title{Plot measles detection method proportions by year}
\usage{
plot_detection_methods(start_year = 2012, end_year = 2025)
}
\arguments{
\item{start_year}{integer year for data to start from}

\item{end_year}{integer year for data to end}
}
\value{
a ggplot object
}
\description{
Creates a stacked bar chart showing the proportion of measles cases
identified by clinical, epidemiologically-linked, and lab-confirmed
detection methods for each year.
}
\examples{
plot_detection_methods(load_data())

}
