% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_me.R
\name{read_me}
\alias{read_me}
\title{Provides context on the functions within the package for easy terminal use.}
\usage{
read_me()
}
\value{
A text paragraph of each function within the package.
}
\description{
Lists the functions available within the package.
}
\examples{
read_me()
}
