% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_case_summary.R
\name{yearly_case_summary}
\alias{yearly_case_summary}
\title{Breaks down yearly measles cases for a single country}
\usage{
yearly_case_summary(cases_year, country = "Madagascar")
}
\arguments{
\item{tibble}{dataset to be filtered}

\item{string}{a country}
}
\value{
tibble
}
\description{
Breaks down yearly measles cases for a single country
}
\examples{
yearly_case_summary(cases_year, coumtry = "United States")

}
