% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_plot.R
\name{create_hex_plot}
\alias{create_hex_plot}
\title{Create a hexbin plot}
\usage{
create_hex_plot(data, x, y, bins = 20, title = NULL, plot_theme = theme_bw())
}
\arguments{
\item{data}{A data frame containing at least two numeric variables.}

\item{x}{The numeric variable on the x-axis.}

\item{y}{The numeric variable on the y-axis.}

\item{bins}{The number of hexagonal bins. The default is 20.}

\item{title}{The plot title. The default is NULL.}

\item{plot_theme}{The ggplot theme to use. The default is theme_bw().
Common other include theme_classic(), theme_minimal(), and theme_void()}
}
\value{
A ggplot object showing a hexbin plot.
}
\description{
This function creates a hexbin plot for any two numeric variables in a data
frame. Axis labels are automatically created from the variable names.
}
\examples{
cases <- load_data()

create_hex_plot(
  data = cases,
  x = year,
  y = measles_total)
}
