% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_country.R
\name{filter_by_country}
\alias{filter_by_country}
\title{Filter measles dataset by country}
\usage{
filter_by_country(data, country)
}
\arguments{
\item{tibble}{dataset}

\item{string}{country to be filtered by}
}
\value{
tibble
}
\description{
Filter measles dataset by country
}
\examples{
filter_by_country(data, "United States")

}
