% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load monthly cases data}
\usage{
load_data()
}
\value{
A tibble containing monthly case data.
}
\description{
Reads the TidyTuesday 2025 monthly cases dataset from GitHub.
}
\examples{
load_data()
}
