% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_monthly_cases.R
\name{plot_monthly_cases}
\alias{plot_monthly_cases}
\title{Create a line plot of measles cases for a country in a given year}
\usage{
plot_monthly_cases(year = 2012, country = "Madagascar")
}
\arguments{
\item{numerical}{year}

\item{string}{country}
}
\value{
ggplot
}
\description{
Create a line plot of measles cases for a country in a given year
}
\examples{
plot_monthly_cases(year = 2018, country = "Madagascar")

}
